+++ apps/connector/global.d.ts
@@ -1,0 +1,7 @@
+import type { Provider } from './src/contents/provider'
+
+declare global {
+  interface Window {
+    ethereum: Provider
+  }
+}